pyproject.toml
[project]
name = "healthcare-mcp-a2a"
version = "0.1.0"
description = "Multi-Agent Patient Care Coordination Platform using MCP + A2A protocols"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # MCP
    "mcp[cli]>=1.26.0",
    # A2A
    "a2a-sdk>=0.2.7",
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Data validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # FHIR
    "fhir.resources>=7.1.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    # Security
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=43.0.0",
    "passlib[bcrypt]>=1.7.4",
    # HTTP
    "httpx>=0.28.0",
    # Observability
    "structlog>=24.4.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    # Data lineage
    "openlineage-python>=1.25.0",
    "networkx>=3.4.0",
    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "faker>=33.0.0",
]
aws = [
    "boto3>=1.35.0",
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "PT",  # flake8-pytest-style
    "SIM", # flake8-simplify
]
ignore = [
    "S101",  # assert used (fine in tests)
    "S105",  # hardcoded password (false positives on config defaults)
    "S106",  # hardcoded password
    "TCH001", # move import into TYPE_CHECKING (conflicts with Pydantic runtime needs)
    "TCH002",
    "TCH003",
    "UP042",  # prefer StrEnum -- we use (str, Enum) for Python 3.11 compat
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "fhir.resources.*",
    "openlineage.*",
    "a2a.*",
    "mcp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "golden_tests"]
addopts = "--cov=src --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "golden_tests/*"]

[tool.coverage.report]
# Phase 1: 55% (lineage emitter, human_review queue, idempotency, resilience not yet exercised)
# Phase 2-3: 70% (MCP servers + agents add coverage for those modules)
# Phase 4: 80% (integration tests cover end-to-end workflows)
fail_under = 55
show_missing = true
